move all dfn finders
